enum cooked

+++ solutions/08_enums/enums2.rs
@@ -6,7 +6,7 @@ struct Point {
 
 #[derive(Debug)]
 enum Message {
-    Resize { width: usize, height: usize },
+    Resize { width: u64, height: u64 },
     Move(Point),
     Echo(String),
     ChangeColor(u8, u8, u8),
